+++ _CfdMesh.py
@@ -20,23 +20,21 @@
 # *                                                                         *
 # ***************************************************************************
 
-__title__ = "_CfdMeshCart"
+__title__ = "_CfdMesh"
 __author__ = "Bernd Hahnebach"
 __url__ = "http://www.freecadweb.org"
 
 
-class _CfdMeshCart():
-    """ Add cut-cell Cartesian specific properties
+class _CfdMesh:
+    """ CFD mesh properties
     """
 
     # they will be used from the task panel too, thus they need to be outside of the __init__
     known_element_dimensions = ['2D', '3D']
-    known_mesh_algorithm_2D = ['Cartesian']
-    known_mesh_algorithm_3D = ['Cartesian']
-    known_mesh_utility = ['cfMesh', 'snappyHexMesh']
+    known_mesh_utility = ['cfMesh', 'snappyHexMesh', 'gmsh']
 
     def __init__(self, obj):
-        self.Type = "CfdMeshCart"
+        self.Type = "CfdMesh"
         self.Object = obj  # keep a ref to the DocObj for nonGui usage
         obj.Proxy = self  # link between App::DocumentObject to  this object
 
@@ -53,15 +51,15 @@ class _CfdMeshCart():
         obj.STLLinearDeflection = 0.1
 
         obj.addProperty("App::PropertyInteger", "NumberCores", "Base", "Number of parallel cores "
-                                                                       "(only applicable when using sHM) ")
+                                                                       "(only applicable when using snappyHexMesh) ")
         obj.NumberCores = 1
 
         obj.addProperty("App::PropertyLink", "Part", "Mesh Parameters", "Part object to mesh")
         obj.Part = None
 
         obj.addProperty("App::PropertyEnumeration", "MeshUtility", "Mesh Parameters",
-                        "Cut-cell Cartesian meshing utilities")
-        obj.MeshUtility = _CfdMeshCart.known_mesh_utility
+                        "Meshing utilities")
+        obj.MeshUtility = _CfdMesh.known_mesh_utility
         obj.MeshUtility = 'cfMesh'
 
         obj.addProperty("App::PropertyLength", "CharacteristicLengthMax", "Mesh Parameters",
@@ -84,16 +82,8 @@ class _CfdMeshCart():
 
         obj.addProperty("App::PropertyEnumeration", "ElementDimension", "Mesh Parameters",
                         "Dimension of mesh elements (Default 3D)")
-        obj.ElementDimension = _CfdMeshCart.known_element_dimensions
+        obj.ElementDimension = _CfdMesh.known_element_dimensions
         obj.ElementDimension = '3D'
-
-        obj.addProperty("App::PropertyEnumeration", "Algorithm2D", "Mesh Parameters", "2D mesh algorithm")
-        obj.Algorithm2D = _CfdMeshCart.known_mesh_algorithm_2D
-        obj.Algorithm2D = 'Cartesian'
-
-        obj.addProperty("App::PropertyEnumeration", "Algorithm3D", "Mesh Parameters", "3D mesh algorithm")
-        obj.Algorithm3D = _CfdMeshCart.known_mesh_algorithm_3D
-        obj.Algorithm3D = 'Cartesian'
 
     def execute(self, obj):
         return
